Switch to Lucid syntax for attributes.
I'm going all-in with the lucid lookalike thing. Side-effect: the types
are (I think) more intuitive now. Alright, so parents are still
confusing due to the `TermParentArg` business, but leaves are more
intuitive.

I'm tempted to scrap the `TermParent` class and require parents to
always provide a list of attributes, like leaves.